Made the missing-items detection more readable
I moved the globbing code outside of the main for loop.
I also changed 'if missing' to 'if not missing'.
All this reduced the indentation depth of the code.
Which is a good thing, as I couldn't keep track of all the 'break'
and 'continue' statements. Over the time the code grew too many
'for' loops :-)